+++ docs/data/base/components/number-field/UnstyledNumberFieldIntroduction/tailwind/index.tsx
@@ -57,14 +57,14 @@ export default function NumberFieldIntroduction() {
 }
 
 const NumberField = React.forwardRef(function NumberField(
-  props: BaseNumberField.RootProps,
+  props: BaseNumberField.NumberFieldRootProps,
   ref: React.ForwardedRef<HTMLDivElement>,
 ) {
   return <BaseNumberField.Root {...props} ref={ref} />;
 });
 
 const NumberFieldGroup = React.forwardRef(function NumberFieldGroup(
-  props: BaseNumberField.GroupProps,
+  props: BaseNumberField.NumberFieldGroupProps,
   ref: React.ForwardedRef<HTMLDivElement>,
 ) {
   return (
@@ -87,7 +87,7 @@ const NumberFieldGroup = React.forwardRef(function NumberFieldGroup(
 });
 
 const NumberFieldScrubArea = React.forwardRef(function NumberFieldScrubArea(
-  props: BaseNumberField.ScrubAreaProps,
+  props: BaseNumberField.NumberFieldScrubAreaProps,
   ref: React.ForwardedRef<HTMLSpanElement>,
 ) {
   return (
@@ -107,7 +107,7 @@ const NumberFieldScrubArea = React.forwardRef(function NumberFieldScrubArea(
 });
 
 const NumberFieldVirtualCursor = React.forwardRef(function NumberFieldVirtualCursor(
-  props: BaseNumberField.ScrubAreaCursorProps,
+  props: BaseNumberField.NumberFieldScrubAreaCursorProps,
   ref: React.ForwardedRef<HTMLSpanElement>,
 ) {
   return (
@@ -127,7 +127,7 @@ const NumberFieldVirtualCursor = React.forwardRef(function NumberFieldVirtualCur
 });
 
 const NumberFieldInput = React.forwardRef(function NumberFieldInput(
-  props: BaseNumberField.InputProps,
+  props: BaseNumberField.NumberFieldInputProps,
   ref: React.ForwardedRef<HTMLInputElement>,
 ) {
   return (
@@ -170,7 +170,7 @@ const buttonStyles = classNames(
 );
 
 const NumberFieldDecrement = React.forwardRef(function NumberFieldDecrement(
-  props: BaseNumberField.DecrementProps,
+  props: BaseNumberField.NumberFieldDecrementProps,
   ref: React.ForwardedRef<HTMLButtonElement>,
 ) {
   return (
@@ -192,7 +192,7 @@ const NumberFieldDecrement = React.forwardRef(function NumberFieldDecrement(
 });
 
 const NumberFieldIncrement = React.forwardRef(function NumberFieldIncrement(
-  props: BaseNumberField.IncrementProps,
+  props: BaseNumberField.NumberFieldIncrementProps,
   ref: React.ForwardedRef<HTMLButtonElement>,
 ) {
   return (
