+++ docs/pages/base-ui/api/number-field-decrement.json
@@ -12,7 +12,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldDecrement",
   "forwardsRefTo": "HTMLButtonElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldDecrement.tsx",
+  "filename": "/packages/mui-base/src/NumberField/Decrement/NumberFieldDecrement.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
