+++ docs/pages/base-ui/api/number-field-group.json
@@ -12,7 +12,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldGroup",
   "forwardsRefTo": "HTMLDivElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldGroup.tsx",
+  "filename": "/packages/mui-base/src/NumberField/Group/NumberFieldGroup.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
