+++ docs/pages/base-ui/api/number-field-increment.json
@@ -12,7 +12,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldIncrement",
   "forwardsRefTo": "HTMLButtonElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldIncrement.tsx",
+  "filename": "/packages/mui-base/src/NumberField/Increment/NumberFieldIncrement.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
