+++ docs/pages/base-ui/api/number-field-input.json
@@ -12,7 +12,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldInput",
   "forwardsRefTo": "HTMLInputElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldInput.tsx",
+  "filename": "/packages/mui-base/src/NumberField/Input/NumberFieldInput.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
