+++ docs/pages/base-ui/api/number-field-scrub-area-cursor.json
@@ -12,7 +12,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldScrubAreaCursor",
   "forwardsRefTo": "HTMLSpanElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldScrubAreaCursor.tsx",
+  "filename": "/packages/mui-base/src/NumberField/ScrubAreaCursor/NumberFieldScrubAreaCursor.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
