+++ docs/pages/base-ui/api/number-field-scrub-area.json
@@ -18,7 +18,7 @@
   "themeDefaultProps": true,
   "muiName": "NumberFieldScrubArea",
   "forwardsRefTo": "HTMLSpanElement",
-  "filename": "/packages/mui-base/src/NumberField/NumberFieldScrubArea.tsx",
+  "filename": "/packages/mui-base/src/NumberField/ScrubArea/NumberFieldScrubArea.tsx",
   "inheritance": null,
   "demos": "<ul><li><a href=\"/base-ui/react-number-field/\">Number Field</a></li></ul>",
   "cssComponent": false
