+++ packages/mui-base/src/NumberField/Decrement/NumberFieldDecrement.types.ts
@@ -1,0 +1,5 @@
+import { NumberFieldRootOwnerState } from '../Root/NumberFieldRoot.types';
+import { BaseUIComponentProps } from '../../utils/BaseUI.types';
+
+export interface NumberFieldDecrementProps
+  extends BaseUIComponentProps<'button', NumberFieldRootOwnerState> {}
