+++ packages/mui-base/src/NumberField/Group/NumberFieldGroup.types.ts
@@ -1,0 +1,5 @@
+import { NumberFieldRootOwnerState } from '../Root/NumberFieldRoot.types';
+import { BaseUIComponentProps } from '../../utils/BaseUI.types';
+
+export interface NumberFieldGroupProps
+  extends BaseUIComponentProps<'div', NumberFieldRootOwnerState> {}
