+++ packages/mui-base/src/NumberField/Increment/NumberFieldIncrement.tsx
@@ -1,17 +1,17 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { useNumberFieldContext } from './NumberFieldContext';
-import type { InputProps } from './NumberField.types';
-import { resolveClassName } from '../utils/resolveClassName';
-import { evaluateRenderProp } from '../utils/evaluateRenderProp';
-import { useRenderPropForkRef } from '../utils/useRenderPropForkRef';
+import type { NumberFieldIncrementProps } from './NumberFieldIncrement.types';
+import { useNumberFieldContext } from '../Root/NumberFieldContext';
+import { resolveClassName } from '../../utils/resolveClassName';
+import { evaluateRenderProp } from '../../utils/evaluateRenderProp';
+import { useRenderPropForkRef } from '../../utils/useRenderPropForkRef';
 
-function defaultRender(props: React.ComponentPropsWithRef<'input'>) {
-  return <input {...props} />;
+function defaultRender(props: React.ComponentPropsWithRef<'button'>) {
+  return <button type="button" {...props} />;
 }
 
 /**
- * The input element for the number field.
+ * The increment stepper button.
  *
  * Demos:
  *
@@ -19,33 +19,29 @@ function defaultRender(props: React.ComponentPropsWithRef<'input'>) {
  *
  * API:
  *
- * - [NumberFieldInput API](https://mui.com/base-ui/react-number-field/components-api/#number-field-input)
+ * - [NumberFieldIncrement API](https://mui.com/base-ui/react-number-field/components-api/#number-field-increment)
  */
-const NumberFieldInput = React.forwardRef(function NumberFieldInput(
-  props: InputProps,
-  forwardedRef: React.ForwardedRef<HTMLInputElement>,
+const NumberFieldIncrement = React.forwardRef(function NumberFieldIncrement(
+  props: NumberFieldIncrementProps,
+  forwardedRef: React.ForwardedRef<HTMLButtonElement>,
 ) {
   const { render: renderProp, className, ...otherProps } = props;
   const render = renderProp ?? defaultRender;
 
-  const { getInputProps, inputRef, inputValue, ownerState } = useNumberFieldContext('Input');
+  const { getIncrementButtonProps, ownerState } = useNumberFieldContext('Increment');
 
-  const mergedInputRef = useRenderPropForkRef(render, forwardedRef, inputRef);
+  const mergedRef = useRenderPropForkRef(render, forwardedRef);
 
-  const inputProps = getInputProps({
-    ref: mergedInputRef,
+  const buttonProps = getIncrementButtonProps({
+    ref: mergedRef,
     className: resolveClassName(className, ownerState),
-    value: inputValue,
-    // If the server's locale does not match the client's locale, the formatting may not match,
-    // causing a hydration mismatch.
-    suppressHydrationWarning: true,
     ...otherProps,
   });
 
-  return evaluateRenderProp(render, inputProps, ownerState);
+  return evaluateRenderProp(render, buttonProps, ownerState);
 });
 
-NumberFieldInput.propTypes /* remove-proptypes */ = {
+NumberFieldIncrement.propTypes /* remove-proptypes */ = {
   // ┌────────────────────────────── Warning ──────────────────────────────┐
   // │ These PropTypes are generated from the TypeScript type definitions. │
   // │ To update them, edit the TypeScript types and run `pnpm proptypes`. │
@@ -64,4 +60,4 @@ NumberFieldInput.propTypes /* remove-proptypes */ = {
   render: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
 } as any;
 
-export { NumberFieldInput };
+export { NumberFieldIncrement };
