+++ packages/mui-base/src/NumberField/Increment/NumberFieldIncrement.types.ts
@@ -1,0 +1,5 @@
+import { NumberFieldRootOwnerState } from '../Root/NumberFieldRoot.types';
+import { BaseUIComponentProps } from '../../utils/BaseUI.types';
+
+export interface NumberFieldIncrementProps
+  extends BaseUIComponentProps<'button', NumberFieldRootOwnerState> {}
