+++ packages/mui-base/src/NumberField/Input/NumberFieldInput.test.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { expect } from 'chai';
 import { act, createRenderer, screen } from '@mui/internal-test-utils';
 import * as NumberField from '@base_ui/react/NumberField';
+import { NumberFieldContext, type NumberFieldContextValue } from '@base_ui/react/NumberField';
 import { fireEvent } from '@testing-library/react';
-import { describeConformance } from '../../test/describeConformance';
-import { NumberFieldContext, NumberFieldContextValue } from './NumberFieldContext';
+import { describeConformance } from '../../../test/describeConformance';
 
 const testContext = {
   getInputProps: (externalProps) => externalProps,
