+++ packages/mui-base/src/NumberField/Input/NumberFieldInput.tsx
@@ -1,17 +1,17 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { useNumberFieldContext } from './NumberFieldContext';
-import type { GroupProps } from './NumberField.types';
-import { resolveClassName } from '../utils/resolveClassName';
-import { evaluateRenderProp } from '../utils/evaluateRenderProp';
-import { useRenderPropForkRef } from '../utils/useRenderPropForkRef';
+import type { NumberFieldInputProps } from './NumberFieldInput.types';
+import { useNumberFieldContext } from '../Root/NumberFieldContext';
+import { resolveClassName } from '../../utils/resolveClassName';
+import { evaluateRenderProp } from '../../utils/evaluateRenderProp';
+import { useRenderPropForkRef } from '../../utils/useRenderPropForkRef';
 
-function defaultRender(props: React.ComponentPropsWithRef<'div'>) {
-  return <div {...props} />;
+function defaultRender(props: React.ComponentPropsWithRef<'input'>) {
+  return <input {...props} />;
 }
 
 /**
- * Groups interactive `NumberField` components together.
+ * The input element for the number field.
  *
  * Demos:
  *
@@ -19,29 +19,33 @@ function defaultRender(props: React.ComponentPropsWithRef<'div'>) {
  *
  * API:
  *
- * - [NumberFieldGroup API](https://mui.com/base-ui/react-number-field/components-api/#number-field-group)
+ * - [NumberFieldInput API](https://mui.com/base-ui/react-number-field/components-api/#number-field-input)
  */
-const NumberFieldGroup = React.forwardRef(function NumberFieldGroup(
-  props: GroupProps,
-  forwardedRef: React.ForwardedRef<HTMLDivElement>,
+const NumberFieldInput = React.forwardRef(function NumberFieldInput(
+  props: NumberFieldInputProps,
+  forwardedRef: React.ForwardedRef<HTMLInputElement>,
 ) {
   const { render: renderProp, className, ...otherProps } = props;
   const render = renderProp ?? defaultRender;
 
-  const { getGroupProps, ownerState } = useNumberFieldContext('Group');
+  const { getInputProps, inputRef, inputValue, ownerState } = useNumberFieldContext('Input');
 
-  const mergedRef = useRenderPropForkRef(render, forwardedRef);
+  const mergedInputRef = useRenderPropForkRef(render, forwardedRef, inputRef);
 
-  const groupProps = getGroupProps({
-    ref: mergedRef,
+  const inputProps = getInputProps({
+    ref: mergedInputRef,
     className: resolveClassName(className, ownerState),
+    value: inputValue,
+    // If the server's locale does not match the client's locale, the formatting may not match,
+    // causing a hydration mismatch.
+    suppressHydrationWarning: true,
     ...otherProps,
   });
 
-  return evaluateRenderProp(render, groupProps, ownerState);
+  return evaluateRenderProp(render, inputProps, ownerState);
 });
 
-NumberFieldGroup.propTypes /* remove-proptypes */ = {
+NumberFieldInput.propTypes /* remove-proptypes */ = {
   // ┌────────────────────────────── Warning ──────────────────────────────┐
   // │ These PropTypes are generated from the TypeScript type definitions. │
   // │ To update them, edit the TypeScript types and run `pnpm proptypes`. │
@@ -60,4 +64,4 @@ NumberFieldGroup.propTypes /* remove-proptypes */ = {
   render: PropTypes.oneOfType([PropTypes.element, PropTypes.func]),
 } as any;
 
-export { NumberFieldGroup };
+export { NumberFieldInput };
