+++ packages/mui-base/src/NumberField/Input/NumberFieldInput.types.ts
@@ -1,0 +1,5 @@
+import { NumberFieldRootOwnerState } from '../Root/NumberFieldRoot.types';
+import { BaseUIComponentProps } from '../../utils/BaseUI.types';
+
+export interface NumberFieldInputProps
+  extends BaseUIComponentProps<'input', NumberFieldRootOwnerState> {}
