+++ packages/mui-base/src/NumberField/ScrubAreaCursor/NumberFieldScrubAreaCursor.types.ts
@@ -1,0 +1,5 @@
+import { NumberFieldRootOwnerState } from '../Root/NumberFieldRoot.types';
+import { BaseUIComponentProps } from '../../utils/BaseUI.types';
+
+export interface NumberFieldScrubAreaCursorProps
+  extends BaseUIComponentProps<'span', NumberFieldRootOwnerState> {}
