+++ packages/mui-base/src/NumberField/index.barrel.ts
@@ -1,0 +1,29 @@
+'use client';
+export { NumberFieldRoot } from './Root/NumberFieldRoot';
+export { NumberFieldContext } from './Root/NumberFieldContext';
+export { useNumberFieldRoot } from './Root/useNumberFieldRoot';
+export type {
+  NumberFieldRootProps,
+  NumberFieldRootOwnerState,
+  NumberFieldContextValue,
+  UseNumberFieldRootParameters,
+  UseNumberFieldRootReturnValue,
+} from './Root/NumberFieldRoot.types';
+
+export { NumberFieldGroup } from './Group/NumberFieldGroup';
+export type { NumberFieldGroupProps } from './Group/NumberFieldGroup.types';
+
+export { NumberFieldInput } from './Input/NumberFieldInput';
+export type { NumberFieldInputProps } from './Input/NumberFieldInput.types';
+
+export { NumberFieldIncrement } from './Increment/NumberFieldIncrement';
+export type { NumberFieldIncrementProps } from './Increment/NumberFieldIncrement.types';
+
+export { NumberFieldDecrement } from './Decrement/NumberFieldDecrement';
+export type { NumberFieldDecrementProps } from './Decrement/NumberFieldDecrement.types';
+
+export { NumberFieldScrubArea } from './ScrubArea/NumberFieldScrubArea';
+export type { NumberFieldScrubAreaProps } from './ScrubArea/NumberFieldScrubArea.types';
+
+export { NumberFieldScrubAreaCursor } from './ScrubAreaCursor/NumberFieldScrubAreaCursor';
+export type { NumberFieldScrubAreaCursorProps } from './ScrubAreaCursor/NumberFieldScrubAreaCursor.types';
