+++ packages/mui-base/src/NumberField/index.ts
@@ -1,10 +1,29 @@
 'use client';
-export { NumberFieldRoot as Root } from './NumberFieldRoot';
-export { NumberFieldGroup as Group } from './NumberFieldGroup';
-export { NumberFieldInput as Input } from './NumberFieldInput';
-export { NumberFieldIncrement as Increment } from './NumberFieldIncrement';
-export { NumberFieldDecrement as Decrement } from './NumberFieldDecrement';
-export { NumberFieldScrubArea as ScrubArea } from './NumberFieldScrubArea';
-export { NumberFieldScrubAreaCursor as ScrubAreaCursor } from './NumberFieldScrubAreaCursor';
+export { NumberFieldRoot as Root } from './Root/NumberFieldRoot';
+export { NumberFieldContext } from './Root/NumberFieldContext';
+export { useNumberFieldRoot } from './Root/useNumberFieldRoot';
+export type {
+  NumberFieldRootProps as RootProps,
+  NumberFieldRootOwnerState as OwnerState,
+  NumberFieldContextValue,
+  UseNumberFieldRootParameters,
+  UseNumberFieldRootReturnValue,
+} from './Root/NumberFieldRoot.types';
 
-export * from './NumberField.types';
+export { NumberFieldGroup as Group } from './Group/NumberFieldGroup';
+export type { NumberFieldGroupProps as GroupProps } from './Group/NumberFieldGroup.types';
+
+export { NumberFieldInput as Input } from './Input/NumberFieldInput';
+export type { NumberFieldInputProps as InputProps } from './Input/NumberFieldInput.types';
+
+export { NumberFieldIncrement as Increment } from './Increment/NumberFieldIncrement';
+export type { NumberFieldIncrementProps as IncrementProps } from './Increment/NumberFieldIncrement.types';
+
+export { NumberFieldDecrement as Decrement } from './Decrement/NumberFieldDecrement';
+export type { NumberFieldDecrementProps as DecrementProps } from './Decrement/NumberFieldDecrement.types';
+
+export { NumberFieldScrubArea as ScrubArea } from './ScrubArea/NumberFieldScrubArea';
+export type { NumberFieldScrubAreaProps as ScrubAreaProps } from './ScrubArea/NumberFieldScrubArea.types';
+
+export { NumberFieldScrubAreaCursor as ScrubAreaCursor } from './ScrubAreaCursor/NumberFieldScrubAreaCursor';
+export type { NumberFieldScrubAreaCursorProps as ScrubAreaCursorProps } from './ScrubAreaCursor/NumberFieldScrubAreaCursor.types';
