+++ packages/mui-base/src/NumberField/utils/constants.ts
