+++ packages/mui-base/src/index.ts
@@ -1,22 +1,4 @@
 export * from './Checkbox/index.barrel';
-export {
-  Root as NumberFieldRoot,
-  Group as NumberFieldGroup,
-  Increment as NumberFieldIncrement,
-  Decrement as NumberFieldDecrement,
-  Input as NumberFieldInput,
-  ScrubArea as NumberFieldScrubArea,
-  ScrubAreaCursor as NumberFieldScrubAreaCursor,
-  type RootProps as NumberFieldRootProps,
-  type GroupProps as NumberFieldGroupProps,
-  type IncrementProps as NumberFieldIncrementProps,
-  type DecrementProps as NumberFieldDecrementProps,
-  type InputProps as NumberFieldInputProps,
-  type ScrubAreaProps as NumberFieldScrubAreaProps,
-  type ScrubAreaCursorProps as NumberFieldScrubAreaCursorProps,
-  type OwnerState as NumberFieldOwnerState,
-} from './NumberField';
+export * from './NumberField/index.barrel';
 export * from './Switch/index.barrel';
 export * from './Tabs/index.barrel';
-
-export * from './useNumberField';
