--- packages/mui-base/src/useNumberField/index.ts
@@ -1,3 +1,0 @@
-'use client';
-export { useNumberField } from './useNumberField';
-export * from './useNumberField.types';
